修复警告

+++ NRShopSugerLib/MyObject/MLPhotoLib/Classes/MLPhotoBrowser/Views/MLPhotoPickerBrowserPhotoScrollView.m
@@ -88,7 +88,7 @@
     }
 }
 
-- (void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(long)buttonIndex{
+- (void)actionSheet:(UIActionSheet *)actionSheet clickedButtonAtIndex:(NSInteger)buttonIndex{
     if (buttonIndex == 0){
         if([UIImagePickerController isSourceTypeAvailable: UIImagePickerControllerSourceTypeSavedPhotosAlbum]) {
